src/python_backend/src/audit_log/requirements.txt
asn1crypto==1.5.1
boto3==1.34.144
botocore==1.34.144
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
filelock==3.15.4
greenlet==3.0.3
idna==3.7
jmespath==1.0.1
packaging==24.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
snowflake-connector-python==3.11.0
snowflake-sqlalchemy==1.6.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.31
tomlkit==0.13.0
typing-extensions==4.12.2
urllib3==1.26.19
